server-ts: Port services/search/expressions/not
