--- src/services/search/expressions/not.js
@@ -1,19 +1,0 @@
-"use strict";
-
-const Expression = require('./expression');
-
-class NotExp extends Expression {
-    constructor(subExpression) {
-        super();
-
-        this.subExpression = subExpression;
-    }
-
-    execute(inputNoteSet, executionContext, searchContext) {
-        const subNoteSet = this.subExpression.execute(inputNoteSet, executionContext, searchContext);
-
-        return inputNoteSet.minus(subNoteSet);
-    }
-}
-
-module.exports = NotExp;
