+++ src/services/search/services/parse.js
@@ -3,7 +3,7 @@
 const dayjs = require("dayjs");
 const AndExp = require('../expressions/and');
 const OrExp = require('../expressions/or.js');
-const NotExp = require('../expressions/not.js');
+const NotExp = require('../expressions/not');
 const ChildOfExp = require('../expressions/child_of');
 const DescendantOfExp = require('../expressions/descendant_of');
 const ParentOfExp = require('../expressions/parent_of.js');
